split tests across multiple files to match files in package
